Clang-format, step 2

+++ code/driver/test/object_file_utils_test.cc
@@ -49,37 +49,34 @@ namespace fs = boost::filesystem;
 namespace insieme {
 namespace driver {
 
-TEST(ObjectFile, HelloWorldTest) {
-	core::NodeManager mgr;
-	
-	fe::ConversionJob job(DRIVER_TEST_DIR "/inputs/hello_world.c");
-	auto unit = job.toIRTranslationUnit(mgr);
-	
-	// save tu to temporary file
-	auto file = fs::unique_path(fs::temp_directory_path() / "tmp%%%%%%%%.o");
-	
-	// save translation unit
-	saveLib(unit, file);
-	
-	// check validity
-	EXPECT_TRUE(fs::exists(file));
-	EXPECT_TRUE(isInsiemeLib(file));
-	
-	// this one should fail ..
-	EXPECT_FALSE(isInsiemeLib(DRIVER_TEST_DIR "/inputs/hello_world.c"));
-	
-	// reload translation unit
-	core::NodeManager mgr2;
-	auto tu = loadLib(mgr2, file);
-	
-	EXPECT_EQ(toString(tu), toString(unit));
-	
-	// cleanup
-	if(fs::exists(file)) {
-		fs::remove(file);
+	TEST(ObjectFile, HelloWorldTest) {
+		core::NodeManager mgr;
+
+		fe::ConversionJob job(DRIVER_TEST_DIR "/inputs/hello_world.c");
+		auto unit = job.toIRTranslationUnit(mgr);
+
+		// save tu to temporary file
+		auto file = fs::unique_path(fs::temp_directory_path() / "tmp%%%%%%%%.o");
+
+		// save translation unit
+		saveLib(unit, file);
+
+		// check validity
+		EXPECT_TRUE(fs::exists(file));
+		EXPECT_TRUE(isInsiemeLib(file));
+
+		// this one should fail ..
+		EXPECT_FALSE(isInsiemeLib(DRIVER_TEST_DIR "/inputs/hello_world.c"));
+
+		// reload translation unit
+		core::NodeManager mgr2;
+		auto tu = loadLib(mgr2, file);
+
+		EXPECT_EQ(toString(tu), toString(unit));
+
+		// cleanup
+		if(fs::exists(file)) { fs::remove(file); }
 	}
-}
 
 } // end namespace driver
 } // end namespace insieme
-
