+++ code/utils/src/compiler/compiler.cpp
@@ -54,11 +54,8 @@ namespace compiler {
 namespace fs = boost::filesystem;
 
 Compiler Compiler::getDefaultC99Compiler() {
-
-	const char *envVar = std::getenv("INSIEME_C_COMPILER");
-	if(envVar == nullptr) {
-		envVar = "gcc";
-	}
+	const char* envVar = std::getenv("INSIEME_C_COMPILER");
+	if(envVar == nullptr) { envVar = "gcc"; }
 	Compiler res(envVar);
 	res.addFlag("-x c");
 	res.addFlag("-Wall");
@@ -67,12 +64,9 @@ Compiler Compiler::getDefaultC99Compiler() {
 }
 
 Compiler Compiler::getDefaultCppCompiler() {
+	const char* envVar = std::getenv("INSIEME_CXX_COMPILER");
+	if(envVar == nullptr) { envVar = "g++"; }
 
-	const char *envVar = std::getenv("INSIEME_CXX_COMPILER");
-	if(envVar == nullptr) {
-		envVar = "g++";
-	}
-	
 	Compiler res(envVar);
 	res.addFlag("-x c++");
 	res.addFlag("-lstdc++");
@@ -104,22 +98,21 @@ Compiler Compiler::getDebugCompiler(const Compiler& base, const string& level) {
 string Compiler::getCommand(const vector<string>& inputFiles, const string& outputFile) const {
 	// build up compiler command
 	std::stringstream cmd;
-	
+
 	// some flags are known to be required to be place before the source file
 	vector<string> before;
 	vector<string> after;
-	
+
 	// split up flags
 	for(auto cur : flags) {
 		// the -x option has to be before the input file
 		if(cur[0] == '-' && cur[1] == 'x') {
 			before.push_back(cur);
-		}
-		else {
+		} else {
 			after.push_back(cur);
 		}
 	}
-	
+
 	cmd << executable;
 	cmd << " " << join(" ", before);
 	cmd << " " << join(" ", inputFiles);
@@ -128,12 +121,10 @@ string Compiler::getCommand(const vector<string>& inputFiles, const string& outp
 	cmd << (libs.getPaths().empty() ? "" : " -L") << join(" -L", libs.getPaths());
 	cmd << (libs.getLibs().empty() ? "" : " -l") << join(" -l", libs.getLibs());
 	cmd << " -o " << outputFile;
-	
+
 	// redirect streams if compilation should be 'silent'
-	if(silent) {
-		cmd << " > /dev/null 2>&1";
-	}
-	
+	if(silent) { cmd << " > /dev/null 2>&1"; }
+
 	return cmd.str();
 }
 
@@ -141,9 +132,7 @@ const vector<string> getDefaultIncludePaths(string cmd) {
 	vector<string> paths;
 	char line[256];
 	FILE* file = popen(cmd.c_str(), "r");
-	if(file == NULL) {
-		return paths;
-	}
+	if(file == NULL) { return paths; }
 	bool capture = false;
 	string input;
 	string startPrefix("#include <...> search starts here:");
@@ -153,26 +142,24 @@ const vector<string> getDefaultIncludePaths(string cmd) {
 		//#include <...> search starts here:
 		if(input.substr(0, startPrefix.length()) == startPrefix) {
 			capture = true;
-			//leave this line out
+			// leave this line out
 			continue;
 		}
-		//End of Search list.
+		// End of Search list.
 		if(input.substr(0, stopPrefix.length()) == stopPrefix) {
 			capture = false;
-			//stop after this line
+			// stop after this line
 			break;
 		}
-		
+
 		if(capture) {
-			input.replace(input.begin(),input.begin()+1,"");
-			input.replace(input.end()-1,input.end(),"/");
+			input.replace(input.begin(), input.begin() + 1, "");
+			input.replace(input.end() - 1, input.end(), "/");
 			paths.push_back(input);
 		}
 	}
 	pclose(file);
-	if(paths.empty()) {
-		std::cerr << "ATTENTION: No default include paths found. Terminal local language has to be set to en_XX.\n";
-	}
+	if(paths.empty()) { std::cerr << "ATTENTION: No default include paths found. Terminal local language has to be set to en_XX.\n"; }
 	return paths;
 }
 
@@ -206,9 +193,7 @@ bool compile(const vector<string>& sourcefile, const string& targetfile, const C
 	string&& cmd = compiler.getCommand(sourcefile, targetfile);
 	LOG(INFO) << "Compiling with: " << cmd << std::endl;
 	int res = system(cmd.c_str());
-	if(res) {
-		std::cerr << "Failure with exit status " << res << std::endl;
-	}
+	if(res) { std::cerr << "Failure with exit status " << res << std::endl; }
 	return res == 0;
 }
 
@@ -221,15 +206,11 @@ bool compile(const string& sourcefile, const string& targetfile, const Compiler&
 
 bool compile(const VirtualPrintable& source, const Compiler& compiler) {
 	string target = compileToBinary(source, compiler);
-	if(target.empty()) {
-		return false;
-	}
-	
+	if(target.empty()) { return false; }
+
 	// delete target file
-	if(boost::filesystem::exists(target)) {
-		boost::filesystem::remove(target);
-	}
-	
+	if(boost::filesystem::exists(target)) { boost::filesystem::remove(target); }
+
 	return true;
 }
 
@@ -237,10 +218,8 @@ string compileToBinary(const VirtualPrintable& source, const Compiler& compiler)
 	// create temporary target file name
 	fs::path targetFile = fs::unique_path(fs::temp_directory_path() / "insieme-trg-%%%%%%%%");
 	LOG(DEBUG) << "Using temporary file " << targetFile << " as a target file for compilation.";
-	
-	if(compileToBinary(source, targetFile.string(), compiler)) {
-		return targetFile.string();
-	}
+
+	if(compileToBinary(source, targetFile.string(), compiler)) { return targetFile.string(); }
 	return string();
 }
 
@@ -249,25 +228,24 @@ bool compileToBinary(const VirtualPrintable& source, const string& targetFile, c
 	// create a temporary source file
 	fs::path sourceFile = fs::unique_path(fs::temp_directory_path() / "insieme-src-%%%%%%%%");
 	LOG(DEBUG) << "Using temporary file " << sourceFile << " as a source file for compilation.";
-	
+
 	// write source to file
 	std::fstream srcFile(sourceFile.string(), std::fstream::out);
 	srcFile << source << "\n";
 	srcFile.close();
-	
+
 	// perform compilation
 	bool success = compile(sourceFile.string(), targetFile, compiler);
-	
+
 	// delete source file - only if compilation was a success
 	if(boost::filesystem::exists(sourceFile)) {
 		if(success) {
 			boost::filesystem::remove(sourceFile);
-		}
-		else {
+		} else {
 			std::cerr << "Offending source code can be found in " << sourceFile << std::endl;
 		}
 	}
-	
+
 	return success;
 }
